re PR tree-optimization/85066 (FAIL: gcc.dg/vect/costmodel/x86_64/costmodel-pr84986.c scan-tree-dump vect "vectorized 1 loops")
2018-03-26  Richard Biener  <rguenther@suse.de>

	PR testsuite/85066
	* gcc.dg/vect/costmodel/x86_64/costmodel-pr84986.c: Use long long
	instead of long.

From-SVN: r258851

+++ gcc/testsuite/gcc.dg/vect/costmodel/x86_64/costmodel-pr84986.c
@@ -1,10 +1,10 @@
 /* { dg-do compile } */
-/* { dg-require-effective-target vect_long } */
+/* { dg-require-effective-target vect_long_long } */
 
 int N;
-long fn1(void) {
+long long fn1(void) {
   short i;
-  long a;
+  long long a;
   i = a = 0;
   while (i < N)
     a -= i++;
